기업정보 삭제, 수정 권한 체크 (#44)
* 기업정보 삭제, 수정 권한 체크

* companyJunit

+++ src/main/java/site/metacoding/miniproject/service/CompanyService.java
@@ -2,6 +2,8 @@ package site.metacoding.miniproject.service;
 
 import java.util.List;
 
+import javax.servlet.http.HttpSession;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -26,6 +28,7 @@ public class CompanyService {
   private final CompanyDao companyDao;
   private final CoCheckDao coCheckDao;
   private final SHA256 sha256;
+  private final HttpSession session;
 
   @Transactional(readOnly = true)
   public CompanySessionUser 로그인(CompanyLoginReqDto companyLoginReqDto) {
@@ -57,34 +60,60 @@ public class CompanyService {
 
   public CompanyDetailRespDto findByCompanyIdToCompanyDetail(Integer companyId) {
     Company companyPS = companyDao.findById(companyId);
-    List<CoCheckRespDto> coCheckList = coCheckDao.findByCompanyId(companyPS.getCompanyId());
-    CompanyDetailRespDto companyDetailRespDto = new CompanyDetailRespDto(companyPS, coCheckList);
-    return companyDetailRespDto;
+    // List<CoCheckRespDto> coCheckList =
+    // coCheckDao.findByCompanyId(companyPS.getCompanyId());
+    // CompanyDetailRespDto companyDetailRespDto = new
+    // CompanyDetailRespDto(companyPS, coCheckList);
+    // return companyDetailRespDto;
+    if (companyPS != null) {
+      List<CoCheckRespDto> coCheckList = coCheckDao.findByCompanyId(companyPS.getCompanyId());
+      CompanyDetailRespDto companyDetailRespDto = new CompanyDetailRespDto(companyPS, coCheckList);
+      return companyDetailRespDto;
+    } else {
+      throw new RuntimeException("해당 " + companyId + "로 상세보기를 할 수 없습니다.");
+    }
+
   }
 
   public CompanyUpdateRespDto updateCompany(Integer companyId, CompanyUpdateReqDto companyUpdateReqDto) {
-    // co_check 값 업데이트
-    coCheckDao.deleteById(companyId);
-    for (Integer jobId : companyUpdateReqDto.getJobIds()) {
-      coCheckDao.insert(companyId, jobId);
-    }
-    List<CoCheckRespDto> jobCheckList = coCheckDao.findByCompanyId(companyId);
-
-    // 회원정보 업데이트
     Company companyPS = companyDao.findById(companyId);
-    companyPS.update(companyUpdateReqDto);
-    companyDao.update(companyPS);
-    return new CompanyUpdateRespDto(companyPS, jobCheckList);
+    if (companyPS == null) {
+      throw new RuntimeException("해당 " + companyId + "로 수정할 수 없습니다.");
+    }
+
+    CompanySessionUser coPrincipal = (CompanySessionUser) session.getAttribute("companySessionUser");
+    System.out.println("디버그 세션 : " + coPrincipal.getCompanyId());
+    if (coPrincipal.getCompanyId().equals(companyPS.getCompanyId())) {
+      // co_check 값 업데이트
+      coCheckDao.deleteById(companyId);
+      for (Integer jobId : companyUpdateReqDto.getJobIds()) {
+        coCheckDao.insert(companyId, jobId);
+      }
+      List<CoCheckRespDto> jobCheckList = coCheckDao.findByCompanyId(companyId);
+
+      // 회원정보 업데이트
+      companyPS.update(companyUpdateReqDto);
+      companyDao.update(companyPS);
+      return new CompanyUpdateRespDto(companyPS, jobCheckList);
+
+    } else {
+      throw new RuntimeException("해당 게시글을 수정할 권한이 없습니다.");
+    }
+
   }
 
   public void deleteCompany(Integer companyId) {
-
     Company companyPS = companyDao.findById(companyId);
-    if (companyPS != null) {
-      companyDao.deleteById(companyId);
-      coCheckDao.deleteById(companyId);
-    } else {
+    if (companyPS == null) {
       throw new RuntimeException("해당 " + companyId + "로 삭제를 할 수 없습니다.");
+    }
+    CompanySessionUser coPrincipal = (CompanySessionUser) session.getAttribute("companySessionUser");
+    System.out.println("디버그 세션 : " + coPrincipal.getCompanyId());
+    if (coPrincipal.getCompanyId().equals(companyPS.getCompanyId())) {
+      companyDao.deleteById(companyPS.getCompanyId());
+      coCheckDao.deleteById(companyPS.getCompanyId());
+    } else {
+      throw new RuntimeException("해당 게시글을 지울 권한이 없습니다.");
     }
   }
